+++ types/layout/dock/Files.d.ts
@@ -1,36 +1,22 @@
-import {App} from "../../../siyuan";
+import {App, IFile} from "../../../siyuan";
 import {Tab} from "../Tab";
 import {Model} from "../Model";
-
-interface IFile {
-    icon: string;
-    name1: string;
-    alias: string;
-    memo: string;
-    bookmark: string;
-    path: string;
-    name: string;
-    titleEmpty?: boolean;
-    hMtime: string;
-    hCtime: string;
-    hSize: string;
-    dueFlashcardCount?: string;
-    newFlashcardCount?: string;
-    flashcardCount?: string;
-    id: string;
-    count: number;
-    subFileCount: number;
-}
 
 export declare class Files extends Model {
     element: HTMLElement;
     parent: Tab;
     private actionsElement;
     closeElement: HTMLElement;
+    lastSelectedElement: Element;
     constructor(options: {
         tab: Tab;
         app: App;
     });
+    private handleMsgCallback;
+    private updateDocInfo;
+    private updateDocActionElement;
+    updateDocActions(): void;
+    private updateItemArrow;
     private genNotebook;
     init(init?: boolean): void;
     private onRemove;
@@ -43,15 +29,16 @@ export declare class Files extends Model {
     private onMove;
     private onLsHTML;
     private onLsSelect;
-    private setCurrent;
-    getLeaf(liElement: Element, notebookId: string): void;
+    setCurrent(target: HTMLElement, isScroll?: boolean): void;
+    getLeaf(liElement: Element, notebookId: string, focusUpdate?: boolean): void;
     selectItem(notebookId: string, filePath: string, data?: {
         files: IFile[];
         box: string;
         path: string;
-    }, setStorage?: boolean): void;
+    }, setStorage?: boolean, isSetCurrent?: boolean): Promise<HTMLElement>;
     private getOpenPaths;
     private genDocAriaLabel;
     private genFileHTML;
     private initMoreMenu;
+    private refreshPublishAccessSwitch;
 }